+++ kernel/src/shell/cmds.rs
@@ -37,3 +37,4 @@ pub mod wait;
 pub mod wipe;
 pub mod write;
 pub mod pci;
+pub mod grep;